[distributed] NCCLflight recorder timeout fix
Because isCompleted() returns true on an exception, a timeout exception
will cause the flight recorder to consider the event completed even though it timed out.

This changes the logic to explicitly query the completion events on "retirement"
when the work item leaves the workMetaList. We mark events as retired so
we can distinguish between an event still in the queue but not completed and one
that timed out.

ghstack-source-id: 660820b54086e63df752d6f4000f5732714be42e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114804