Refactor can_auto_functionalize
In preparation for the next PR up in the stack, which is going to update
"can_auto_functionalize" to support more operators than just ones that
return nothing. We are unable to auto-generate FakeTensor kernels for
operators that do not return nothing, but we are able to generate
functionalization kernels for operators that return something.

Test Plan:
Existing tests

ghstack-source-id: 70b847c33bba2aec973303ed8c116b37f4227a86
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115134