[C10D] Use future for flight recorder dump
Replaces the "always sleep 30 sec before abort" with "wait up to 30 sec
for the future to complete then abort".  The difference in this case is
the abort happens as soon as the dump finishes up to a maximum, instead
of always waiting the maximum.

Simplifies the test case for dump on timeout, which can now wait for
rank0 abort and assume the output file will be ready.

ghstack-source-id: f7ba826f2955655afbcf089655ebbaea118e2865
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115176