[C10D] Add duration_ms to flight recorder
Measures the duration of a collective operation using nccl start/end
events and includes this duration (in ms) in the flight recorder data.

duration_ms will be an optional field, since it only works when
timing is enabled.  Currently timing is enabled when flight recorder
is enabled, but this is not a strict requirement.  Duration is also
not available for collectives not in a completed state.

ghstack-source-id: 45df4b5a2727f6b237ff30e46cf36c494cb3dc9c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114817